requirements.txt
playwright
playwright-stealth
fastapi
uvicorn
pyyaml
